SFORM-95 各DAOのdeleteメソッドのレスポンス型をBigInt->Intに

+++ domain/src/main/scala/net/macolabo/sform2/domain/models/daos/TransferConfigDAO.scala
@@ -46,7 +46,7 @@ trait TransferConfigDAO {
    * @param session          DB Session
    * @return TransferConfig
    */
-  def delete(userGroup: String, transferConfigId: BigInt)(implicit session: DBSession = autoSession): BigInt
+  def delete(userGroup: String, transferConfigId: BigInt)(implicit session: DBSession = autoSession): Int
 
 }
