+++ domain/src/main/scala/net/macolabo/sform2/domain/models/daos/TransferConfigDAOImpl.scala
@@ -114,7 +114,7 @@ class TransferConfigDAOImpl extends TransferConfigDAO {
    * @param session          DB Session
    * @return TransferConfig
    */
-  def delete(userGroup: String, transferConfigId: BigInt)(implicit session: DBSession = autoSession): BigInt = {
+  def delete(userGroup: String, transferConfigId: BigInt)(implicit session: DBSession = autoSession): Int = {
     withSQL {
       val c = TransferConfig.column
       deleteFrom(TransferConfig)
