+++ domain/src/main/scala/net/macolabo/sform2/domain/models/daos/TransferConfigMailAddressDAO.scala
@@ -52,7 +52,7 @@ trait TransferConfigMailAddressDAO {
    * @param session                     DB Session
    * @return Result
    */
-  def delete(userGroup: String, transferConfigMailAddressId: BigInt)(implicit session: DBSession = autoSession): BigInt
+  def delete(userGroup: String, transferConfigMailAddressId: BigInt)(implicit session: DBSession = autoSession): Int
 }
