+++ domain/src/main/scala/net/macolabo/sform2/domain/models/daos/TransferConfigSalesforceObjectDAOImpl.scala
@@ -89,7 +89,7 @@ class TransferConfigSalesforceObjectDAOImpl extends TransferConfigSalesforceObje
    * @param session DB Session
    * @return Result
    */
-  def delete(userGroup: String, transferConfigSalesforceObjectId: BigInt)(implicit session: DBSession = autoSession): BigInt = {
+  def delete(userGroup: String, transferConfigSalesforceObjectId: BigInt)(implicit session: DBSession = autoSession): Int = {
     withSQL {
       val c = TransferConfigSalesforceObject.column
       deleteFrom(TransferConfigSalesforceObject)
