+++ domain/src/main/scala/net/macolabo/sform2/domain/models/daos/TransferConfigSalesforceObjectFieldDAO.scala
@@ -42,5 +42,5 @@ trait TransferConfigSalesforceObjectFieldDAO {
    * @param session                               DB Session
    * @return Result
    */
-  def delete(userGroup: String, transferConfigSalesforceObjectFieldId: BigInt)(implicit session: DBSession = autoSession): BigInt
+  def delete(userGroup: String, transferConfigSalesforceObjectFieldId: BigInt)(implicit session: DBSession = autoSession): Int
 }
